source/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt source/requirements.txt
#
ablog==0.10.29
    # via -r source/requirements.txt
aiohttp==3.8.1
    # via -r source/requirements.txt
aiosignal==1.2.0
    # via
    #   -r source/requirements.txt
    #   aiohttp
alabaster==0.7.12
    # via
    #   -r source/requirements.txt
    #   sphinx
async-timeout==4.0.2
    # via
    #   -r source/requirements.txt
    #   aiohttp
attrs==22.1.0
    # via
    #   -r source/requirements.txt
    #   aiohttp
babel==2.10.3
    # via
    #   -r source/requirements.txt
    #   sphinx
beautifulsoup4==4.11.1
    # via
    #   -r source/requirements.txt
    #   pydata-sphinx-theme
certifi==2022.9.14
    # via
    #   -r source/requirements.txt
    #   requests
charset-normalizer==2.1.1
    # via
    #   -r source/requirements.txt
    #   aiohttp
    #   requests
colorama==0.4.5
    # via
    #   -r source/requirements.txt
    #   sphinx-autobuild
docutils==0.17.1
    # via
    #   -r source/requirements.txt
    #   ablog
    #   myst-parser
    #   pybtex-docutils
    #   pydata-sphinx-theme
    #   sphinx
    #   sphinx-togglebutton
    #   sphinxcontrib-bibtex
feedgen==0.9.0
    # via
    #   -r source/requirements.txt
    #   ablog
frozenlist==1.3.1
    # via
    #   -r source/requirements.txt
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   -r source/requirements.txt
    #   requests
    #   yarl
imagesize==1.4.1
    # via
    #   -r source/requirements.txt
    #   sphinx
invoke==1.7.1
    # via
    #   -r source/requirements.txt
    #   ablog
jinja2==3.1.2
    # via
    #   -r source/requirements.txt
    #   myst-parser
    #   sphinx
latexcodec==2.0.1
    # via
    #   -r source/requirements.txt
    #   pybtex
linkify-it-py==2.0.0
    # via -r source/requirements.txt
livereload==2.6.3
    # via
    #   -r source/requirements.txt
    #   sphinx-autobuild
lxml==4.9.1
    # via
    #   -r source/requirements.txt
    #   feedgen
markdown-it-py==2.1.0
    # via
    #   -r source/requirements.txt
    #   mdit-py-plugins
    #   myst-parser
markupsafe==2.1.1
    # via
    #   -r source/requirements.txt
    #   jinja2
mdit-py-plugins==0.3.0
    # via
    #   -r source/requirements.txt
    #   myst-parser
mdurl==0.1.2
    # via
    #   -r source/requirements.txt
    #   markdown-it-py
multidict==6.0.2
    # via
    #   -r source/requirements.txt
    #   aiohttp
    #   yarl
myst-parser==0.18.0
    # via -r source/requirements.txt
packaging==21.3
    # via
    #   -r source/requirements.txt
    #   pydata-sphinx-theme
    #   sphinx
pybtex==0.24.0
    # via
    #   -r source/requirements.txt
    #   pybtex-docutils
    #   sphinxcontrib-bibtex
pybtex-docutils==1.0.2
    # via
    #   -r source/requirements.txt
    #   sphinxcontrib-bibtex
pydata-sphinx-theme==0.8.1
    # via
    #   -r source/requirements.txt
    #   sphinx-book-theme
pygments==2.13.0
    # via
    #   -r source/requirements.txt
    #   sphinx
pyparsing==3.0.9
    # via
    #   -r source/requirements.txt
    #   packaging
python-dateutil==2.8.2
    # via
    #   -r source/requirements.txt
    #   ablog
    #   feedgen
pytz==2022.2.1
    # via
    #   -r source/requirements.txt
    #   babel
pyyaml==6.0
    # via
    #   -r source/requirements.txt
    #   myst-parser
    #   pybtex
    #   sphinx-book-theme
requests==2.28.1
    # via
    #   -r source/requirements.txt
    #   sphinx
    #   sphinxcontrib-youtube
six==1.16.0
    # via
    #   -r source/requirements.txt
    #   latexcodec
    #   livereload
    #   pybtex
    #   python-dateutil
snowballstemmer==2.2.0
    # via
    #   -r source/requirements.txt
    #   sphinx
soupsieve==2.3.2.post1
    # via
    #   -r source/requirements.txt
    #   beautifulsoup4
sphinx==4.5.0
    # via
    #   -r source/requirements.txt
    #   ablog
    #   myst-parser
    #   pydata-sphinx-theme
    #   sphinx-autobuild
    #   sphinx-book-theme
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-thebe
    #   sphinx-togglebutton
    #   sphinxcontrib-bibtex
    #   sphinxcontrib-youtube
    #   sphinxext-opengraph
sphinx-autobuild==2021.3.14
    # via -r source/requirements.txt
sphinx-book-theme==0.3.3
    # via -r source/requirements.txt
sphinx-copybutton==0.5.0
    # via -r source/requirements.txt
sphinx-design==0.3.0
    # via -r source/requirements.txt
sphinx-thebe==0.1.2
    # via -r source/requirements.txt
sphinx-togglebutton==0.3.2
    # via -r source/requirements.txt
sphinxcontrib-applehelp==1.0.2
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-bibtex==2.5.0
    # via -r source/requirements.txt
sphinxcontrib-devhelp==1.0.2
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-jsmath==1.0.1
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-qthelp==1.0.3
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via
    #   -r source/requirements.txt
    #   sphinx
sphinxcontrib-youtube==1.2.0
    # via -r source/requirements.txt
sphinxext-opengraph==0.6.3
    # via -r source/requirements.txt
tornado==6.2
    # via
    #   -r source/requirements.txt
    #   livereload
typing-extensions==4.3.0
    # via
    #   -r source/requirements.txt
    #   myst-parser
uc-micro-py==1.0.1
    # via
    #   -r source/requirements.txt
    #   linkify-it-py
urllib3==1.26.12
    # via
    #   -r source/requirements.txt
    #   requests
watchdog==2.1.9
    # via
    #   -r source/requirements.txt
    #   ablog
wheel==0.37.1
    # via
    #   -r source/requirements.txt
    #   sphinx-togglebutton
yarl==1.8.1
    # via
    #   -r source/requirements.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
